Create chapter_6
Write a program to answer this question. Look at Example 4 in chapter 6.5. Suppose a cookie shop has four different kinds of cookies. How many different ways can six cookies be chosen? Assume the type of cookie and not the individual cookies or the order in which they are chosen matters.