Create spot_v0.5
Note: You will have to create your own Spotify developer account and get your own client id and secret to use this. It's just a prototype currently and I plan to turn it into a simple command shortcut for Spotify in case you're playing a game or something and want to turn on music.